Refactoring and a new pregen message
