+++ core/src/main/java/com/volmit/iris/core/safeguard/IrisSafeguard.java
@@ -1,4 +1,4 @@
-package com.volmit.iris.engine.safeguard;
+package com.volmit.iris.core.safeguard;
 
 import com.volmit.iris.Iris;
