+++ core/src/main/java/com/volmit/iris/core/safeguard/PerformanceSFG.java
@@ -1,4 +1,4 @@
-package com.volmit.iris.engine.safeguard;
+package com.volmit.iris.core.safeguard;
 
 import com.volmit.iris.core.IrisSettings;
 import oshi.SystemInfo;
